+++ tests/Order/OrderServiceTest.php
@@ -16,6 +16,7 @@ use Wizaplace\SDK\Exception\OrderNotFound;
 use Wizaplace\SDK\Order\AfterSalesServiceRequest;
 use Wizaplace\SDK\Order\CreateOrderReturn;
 use Wizaplace\SDK\Order\OrderCommitmentCommand;
+use Wizaplace\SDK\Order\OrderItem;
 use Wizaplace\SDK\Order\OrderReturnStatus;
 use Wizaplace\SDK\Order\OrderService;
 use Wizaplace\SDK\Order\OrderStatus;
@@ -175,7 +176,7 @@ final class OrderServiceTest extends ApiTestCase
         $this->assertSame(1.4, $order->getTaxtotal());
         $this->assertEquals(OrderStatus::COMPLETED(), $order->getStatus());
         $this->assertSame('TNT Express', $order->getShippingName());
-        $this->assertCount(1, $order->getOrderItems());
+        $this->assertCount(2, $order->getOrderItems());
         $this->assertSame('Please deliver at the front desk of my company.', $order->getCustomerComment());
 
         // Premier orderItem
@@ -201,6 +202,18 @@ final class OrderServiceTest extends ApiTestCase
         $this->assertSame('The World Company Inc.', $orders[3]->getCompanyName());
 
         return $orders;
+    }
+
+    public function testGetOrdersWhichReturnsProductImageId()
+    {
+        $apiClient = $this->buildApiClient();
+        $apiClient->authenticate('customer-1@world-company.com', 'password-customer-1');
+        $orderService = new OrderService($apiClient);
+        $order1 = $orderService->getOrder(1);
+        $order2 = $orderService->getOrder(4);
+
+        $this->assertSame('0', $order1->getOrderItems()[0]->getProductImageId());
+        $this->assertSame('0', $order2->getOrderItems()[1]->getProductImageId());
     }
 
     public function testCommitOrder()